+++ src/plugins/oldwebsocket.js
@@ -1,5 +1,0 @@
-import OldWSConnection from '../transports/old-websocket';
-
-export default function(client) {
-    client.transports['old-websocket'] = OldWSConnection;
-}